Refactor success alert for updating permissions
We'll shortly be adding success alerts for adding and removing access,
so being able to pass in message and description in the controller makes
this simple.

We need to include `ActionView::Helpers::TagHelper` in the helper here
to make the `tag` method available to be called in controllers. We could
include it directly in the controller but including it in the helper
itself means we're less likely to forget this.

Unforunately we can't call `serialize` in the helper/controller as the
escaping is stripped out when it's serialized as JSON when stored as a
browser cookie. [1]

This same serialization is what converts symbols to strings, meaning we
also can't pass `flash[:success_alert]` to the Success Alert, as the
`message` and `description` keys are converted to strings to store in
the cookie, while the component expected an object with symbol keys.[2]

[1]: https://groups.google.com/g/rubyonrails-core/c/z52zgDgUmbs
[2]: bigbinary.com/blog/flash-access-changes-rails-4-1