Keep email and password input widths equal
The show/hide button cuts into the space for the password input. In
order for the email and password inputs to stay the same width, we need
to add padding to the email input's wrapper equal to the space taken up
by the show/hide button (including margin)

We only want this change to apply to the signin form, not anywhere we
have an email input, like on the create/invite new user page. I've used
the form action in the selector to make this specific, which is a bit
unconventional - we could add and use a class instead, but this works
without changing the template. It is however reliant on the Device new
session path not changing