Switch back to Jasmine unit tests
When improving the clear button for the accessible autocomplete
component (0c5d5a1), we moved to integration tests using Capybara with
Selenium, instead of unit tests with Jasmine. This was because the
button was moved outside of the autocomplete module, which caused issues
with Jasmine-based testing

This slightly restructures the template where we use accessible
autocomplete to wrap all the elements the JavaScript needs to access in
the `div[data-module=accessible-autocomplete]` element. We then use
dependency injection to pass the module to custom functions that enhance
the form. These changes allow us to revert to the unit test approach,
which in turn means we can introduce the autocomplete component into
other templates without needing to test context-independent
functionality in multiple integration test specs

I've retained a couple of the JavaScript-enabled integration tests where
they're more about checking everything works in a specific context