Refactor Integration tests for updating apps with many permissions
The tests in this file kept this DRY but were a little tricky to figure
out at a glance what each test was doing. We've refactored them a fair
bit to remove some of the abstractions and make them easier to read,
while still looping over the two use cases of updating one's own
permissions and updating those of another user.

This refactor was prompted by the addition of a new Accessible
Autocomplete component to the Organisations page. We've added a new
AutocompleteHelper in this commit to perform some simple Autocomplete
actions and prevent us needing to re-write a few lines of code to find
and select an option.

It's used here for updating permissions tests, and will be used in the
next commit in inviting users tests when adding autocomplete to the
organisation field.

Apologies for multiple changes going on in this commit, it's
essentially:
- Adding a new Autocomplete Helper for reuse in other areas of the
  service that use the Accessible Autocomplete.
- Refactoring the tests to use this new helper and to improve
  readability.

Co-authored-by: George Eaton <george@dxw.com>