Consolidate organisation options methods
These are essentially doing the same thing, so there's no need to have
multiple methods for them

I've deleted a test from a controller spec that duplicates a helper test
added here

I had considered removing the `Organisation` policy scope. The scope
gives GOV.UK admins access to all organisations and everyone else access
to none, but is only used in places that are accessible only to GOV.UK
admins (also based on the organisation policy), so it felt like a bit of
a double-authorisation situation. However, George and I felt a bit
uneasy about the potential for the `options_for_organisation_select`
method being used elsewhere in future, and the authorisation therefore
being skipped. As a result, we've retained the double authorisation, and
I've included a scope-based context in the helper tests

Another thing I spotted is that the `User` model has a
`manageable_organisations` method, in turn relying on e.g.
`Roles::Admin.manageable_organisations_for` or
`Roles::SuperOrganisationAdmin.manageable_organisations_for`. This is
used for filtering users by organisation in the users page/table, rather
than populating a select field when creating/editing users, but the name
makes the remits a little hard to differentiate without close inspection