Use `Organisation#name_with_abbreviation` method
This wasn't being used in the organisations index. We had custom logic
which resulted in organisations without abbreviations having empty
brackets after their name, which looked odd. This method separates the
name and abbreviation with a dash rather than putting the abbreviation
in brackets, but there doesn't seem to be an obvious reason why both
uses can't be the same

The dash used in the method was a non-standard character, so that's
fixed here