Split organisation index into two tabs
... one for active organisations, one for inactive. This will reduce the
closed organisation noise and make it easier to look through active
organisations by default. We'll separately set a default order on
organisations so that closed ones come last, but this felt like a nicer
way to handle the organisations index, and follows the pattern
established by the applications index

This required a change to a 2SV integration test so that it looks for a
table within the active tab/container, since both tabs and hence both
tables are within the underlying HTML. Initially I looked to extend
existing methods in the `HtmlTableHelpers` file to allow us to look for
a table within the right container, but I noticed that the helper's
methods were only used in this integration test. Since their use case
was specific to this spec, I've opted to reduce the abstraction here and
keep it simple, similar to recent minor shifts in the approach to access
and permissions integration tests. This reduces complexity in the
codebase, and brings all the relevant code closer to the context of the
tests - I'm not always a proponent of this approach, but it felt
appropriate here given the limited use case!