Link to access and permissions public page
We had a question about the meaning of some of the content on this page,
particularly the definition of a publishing manager. We wrote
documentation about this in this repo's docs, which it turns out are
automatically published on gov.uk in the dev docs. We're therefore
linking to this resource for readers who want to know more

The existing paragraphs are largely - if not fully - duplicated from
those docs, which doesn't create some extra maintenance burden vs having
one source of truth, but it's probably useful to have at least a short
explainer here without having to click through to a linked external (if
still gov.uk) page