Update unsuspension guidance
We have a rule in the code that a user must log in within seven days of
being unsuspended, or the suspension will be reenforced. This doesn't
seem to be documented outside of the code (at least not anywhere
obvious). We had a support request come in about a user affected by
this, so we should clarify this requirement on the unsuspension form