Add rake task to mark an organisation as closed
The `OrganisationFetcher` does not handle the situation of an
organisation having been deleted in Whitehall without first being
closed.

This means a deleted organisation remains marked as open, and can have
users assigned to it in Signon.

Adding a rake task to manually mark organisations as closed, in the
event this occurs again, so the organisation no longer appears as an
option to which users can be assigned in the user interface.