Add rake task to move all users from one organisation to another
At the moment, we can only bulk move users from one organisation to
another by uploading a CSV file with their email addresses.

There may be circumstances in which we need to move all the users from
one department to another. For example the motivation to add this task
is that some users have been incorrectly assigned to an older, deleted
instance of an organisation.

Adding a rake task to enable us to do this without needing a CSV file.