Skip API authentication in dev environment
Now we’re using this API locally in a couple of applications, the apps
are falling over when trying to access the endpoint. I looked into
seeding an application user, but this is tricky because we can’t
guarantee what the token will be very easily. I’ve worked around this by
skipping the authentication if we’re in the development environment.

It’s not ideal, but saves a lot of messy config in the Docker containers
to get apps that use the Signon API up and running.