Raise custom error with overlong request paths
We often see `ActionDispatch::Cookies::CookieOverflow` errors caused by
requests being made with really long paths (via a URL query). Earlier
investigation suggested these are likely (failed) malicious requests,
e.g. "/?unix:" followed by "A" repeated a few thousand times then a pipe
and a dangerous-looking URL

We want to stay aware of genuine cookie overflow issues in case they're
indiciative of architectural/code design issues that might be breaking
intended behaviour, so blanket suppression of the error in Sentry
doesn't feel like a good option

From fresh investigation it looks like we raise
`ActionDispatch::Cookies::CookieOverflow` errors when the full path's
size in bytes is over half the maximum cookie size. I've confirmed this
locally and in integration by gradually reducing the length of the
query. The maximum cookie size is consistently 4096 bytes (4kB) across
all environments, so it's possible that the threshold is 2048
rather than half the max, but for now there's no difference

The exception occurs while processing the `authenticate_user!`
`before_action`. With this change we now have a custom method which
raises a custom error if the path is too long before calling `super`. It
would be nice to perform the check only if and when the cookie overflow
error is raised, but I wasn't able to rescue from this after calling
`super`. We'll be able to suppress just this custom error by explicitly
excluding it from reporting with `GovukError` (and retain other cookie
overflow error reporting)

---

This was a good signpost: https://github.com/heartcombo/devise/pull/5345