Don't set overlong return to paths in cookie
This is similar to a recent change in gds-sso:
https://github.com/alphagov/gds-sso/commit/c07105927a014f9468fb5f03730e219b2358cc69

We introduced that change after seeing
`ActionDispatch::Cookies::CookieOverflow` errors reported to Sentry from
Authenticating Proxy. We occasionally see the same errors in Signon. We
aren't certain that an overlong 'return to' path is causing the issue,
but it is a known variable-length value that gets set in the
session cookie, which is the location of the overflow. The path gets set
in the cookie when a user tries to access a page but gets redirected to
sign in. They get redirected to the return to path on successful
authentication

The approach here is very similar to what we did in gds-sso, though here
we're introducing a custom Warden failure app for the first time. We're
also using the `ActionDispatch::Cookies::MAX_COOKIE_SIZE` constant to
get the seemingly magic number limit for 'okay' return to path lengths -
since gds-sso is used in both Rails and non-Rails apps, we couldn't use
this Rails constant there

Relevant parts of the Devise code:
- https://github.com/heartcombo/devise/blob/cf93de390a29654620fdf7ac07b4794eb95171d0/lib/devise/failure_app.rb#L244-L250
- https://github.com/heartcombo/devise/blob/cf93de390a29654620fdf7ac07b4794eb95171d0/lib/devise/controllers/store_location.rb#L36-L41
- https://github.com/heartcombo/devise/blob/cf93de390a29654620fdf7ac07b4794eb95171d0/lib/devise/controllers/store_location.rb#L51-L54

We considered the following alternative approaches, but opted for the
failure app option since it makes use of public Devise configuration
even if we're overriding the default behaviour of a protected method:
- defining a custom `authenticate_user!` method in the `RootController`
  and raising a custom error if the `request.fullpath.bytesize` is too
  long, then adding this error to the excluded exceptions in the
  `GovukError` config
- monkey patching the private method
  `ActionDispatch::Session::CookieStore#set_cookie` so that we scrub
  overlong `user_return_to` values (resetting to root) before performing
  the overflow check (in
  `ActionDispatch::Cookies::SerializedCookieJars#check_for_overflow!`)
- monkey patching
  `Devise::Controllers::StoreLocation#store_location_for` so we only
  remember the path when its length is below a particular threshold

We write a path to the cookie in a few other places, but we believe the
errors we're seeing are from unauthenticated requests which wouldn't
reach these parts of the app:
- `SessionsController#create` via `store_location_for` (a Devise method)
- `TwoStepVerificationHelper#handle_failed_second_step` (via
  `#handle_two_step_verification`)