Call load-analytics and populate more meta tags
Without the inclusion of the load-analytics script, we don't actually
fire the page view events (as of June 2025). So we have to include this
now to actually get any meaningful analytics into GA4.

While we're here, we may as well populate the other govuk: meta tags
that whitehall does. These are almost the same as
https://github.com/alphagov/whitehall/blob/main/app/views/layouts/design_system.html.erb#L5-L10
but because Signon doesn't use gds-sso it's `role_name` instead of
`role` (and I've used tag.meta instead of interpolating <metas> because
I think it shows more flair).