Only send the token with longest validity to Kubernetes
At the moment, we iterate through the tokens for each application and
submit it to the Kubernetes secret manager with the same name, meaning
we overwrite the previous each time.

Since only one token can exist per application in the secrets manager
(as each time we send, we use the same secret name), we are relying on
the order of tokens returned by the database being the order of longest
validity, as we iterate through the database results and overwrite the
previous each time.

As the order is not specified, we are assuming the database will
return the records in order of ID. Whilst this usually happens, it is
not guaranteed.

Updating to send only the token with the longest validity. This will
mean we always use the newly generated token and won't need to iterate
over submitting old tokens that we don't want to use.